feat: update post and like models, enhance API responses, and improve error handling

+++ src/app.service.ts
@@ -1,6 +1,8 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
+  Logger,
   NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
@@ -20,13 +22,15 @@ export class AppService {
     private view: ViewService,
   ) {}
 
+  private readonly logger = new Logger(PostService.name);
+
   async create_post(uid: string, data: { content: string; media: string[] }) {
     const user = await this.user.user({ where: { user_id: uid } });
     if (!user) throw new NotFoundException();
     const post = this.post.create({
       author_id: uid,
       author_username: user.name,
-      author_avatar: user.avatar,
+      author_avatar: user.avatar ?? undefined,
       content: data.content,
       media: data.media,
     });
@@ -41,9 +45,7 @@ export class AppService {
   ) {
     const user = await this.user.user({ where: { user_id: uid } });
     if (!user) throw new NotFoundException();
-    const parent = await this.post.post({
-      id: pid,
-    });
+    const parent = await this.post.post({ _id: pid });
     if (!parent) throw new NotFoundException('Post not found');
 
     const post = await this.post.create({
@@ -56,6 +58,11 @@ export class AppService {
     });
 
     if (!post) throw new ForbiddenException('Something went wrong');
+
+    await this.comment.create({
+      parent_post: pid,
+      child_post: post.id as string,
+    });
     parent.child_count++;
     await parent.save();
 
@@ -63,10 +70,20 @@ export class AppService {
   }
 
   async like_post(uid: string, pid: string) {
-    const post = await this.post.post({
-      id: pid,
-    });
+    // try {
+    const post = await this.post.post({ _id: pid }).lean();
     if (!post) throw new NotFoundException();
+
+    const user_like = await this.like.like({
+      where: {
+        post_id_user_id: {
+          post_id: pid,
+          user_id: uid,
+        },
+      },
+    });
+
+    if (user_like) throw new BadRequestException('User already liked post');
 
     await this.like.create({
       user: { connect: { user_id: uid } },
@@ -75,13 +92,26 @@ export class AppService {
     post.like_count++;
     await post.save();
     return post;
+    // } catch (error) {
+    //   this.logger.error('Error creating post', error);
+    //   throw new InternalServerErrorException('Unable to create post');
+    // }
   }
 
   async unlike_post(uid: string, pid: string) {
-    const post = await this.post.post({
-      id: pid,
-    });
+    const post = await this.post.post({ _id: pid }).lean();
     if (!post) throw new NotFoundException();
+
+    const user_like = await this.like.like({
+      where: {
+        post_id_user_id: {
+          post_id: pid,
+          user_id: uid,
+        },
+      },
+    });
+
+    if (!user_like) throw new BadRequestException("User hasn't liked post yet");
 
     await this.like.remove({
       post_id_user_id: {
@@ -94,14 +124,20 @@ export class AppService {
     return post;
   }
 
-  get_post(pid: string) {
-    return this.post.post({ _id: pid });
+  async get_post(pid: string) {
+    const post = await this.post.post({ _id: pid }).lean();
+    if (!post) throw new NotFoundException('Post not found');
+    post.views_count++;
+    await post.save();
+    return post;
   }
 
   get_user_posts(aid: string) {
-    return this.post.posts({
-      author_id: aid,
-    });
+    return this.post
+      .posts({
+        author_id: aid,
+      })
+      .lean();
   }
 
   async get_children(pid: string) {
@@ -109,13 +145,9 @@ export class AppService {
       where: { parent_post: pid },
     });
 
-    const posts: string[] = [];
-    comments.forEach((comment) => {
-      posts.push(comment.child_post);
-    });
-    // const children.
+    const postsIds = comments.map((comment) => comment.child_post);
     return this.post.posts({
-      _id: { in: posts },
+      _id: { $in: postsIds },
     });
   }
 
@@ -130,18 +162,15 @@ export class AppService {
       throw new UnauthorizedException('Only Owner of Post can edit post');
 
     await post.updateOne({ data });
+    await post.save();
     return post;
-    // return this.post.
   }
 
-  async delete_post(uid, pid) {
-    const post = await this.post.post({
-      author_id: uid,
-      _id: pid,
-    });
-
+  async delete_post(uid: string, pid: string) {
+    const post = await this.post.post({ _id: pid });
     if (!post) throw new NotFoundException('Post not found for user');
-
+    if (post.author_id !== uid)
+      throw new UnauthorizedException('Not Owner of Post');
     await post.deleteOne();
     return post;
   }
